+++ app/401/page.tsx
@@ -1,0 +1,3 @@
+export default function ErrorPage() {
+	return <p>401: Forbidden</p>;
+}
